Apply auth to blog entry deletion

+++ part4/bloglist/controllers/blogs.js
@@ -11,8 +11,6 @@ blogRouter.get('/', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token || !decodedToken.id)
-    return response.status(401).json({ error: 'token missing or invalid' })
   const user = await User.findById(decodedToken.id)
 
   const body = request.body
@@ -43,6 +41,12 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+
+  if (!user.blogs.includes(request.params.id))
+    return response.status(401).json({ error: 'Permission denied' })
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
